Refactor bucket iteration
